Revert "[FE] concurrently 설치, 각 페이지 기능 구현, 리팩토링 (#13)"
This reverts commit 656d88d87da5dbf5e91dd3c4d12aa3abfeaf7c0d.
