+++ client/package.json
@@ -10,7 +10,6 @@
     "buildcss": "tailwindcss build -i ./src/index.css -o ./dist/index.css --watch"
   },
   "dependencies": {
-    "dotenv": "^16.0.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.7.1",
