--- client/src/components/Background.jsx
@@ -1,4 +1,0 @@
-import tw from 'tailwind-styled-components';
-
-const Background = tw.div`relative mx-auto flex h-screen w-screen max-w-[480px] flex-col overflow-hidden bg-paper bg-cover bg-no-repeat`;
-export { Background };
