--- client/src/components/cover/index.js
@@ -1,2 +1,0 @@
-export * from './Main';
-export * from './PasswordModal';
