--- client/src/components/homepage/index.js
@@ -1,5 +1,0 @@
-export * from './Dot';
-export * from './MemberInfo';
-export * from './Header';
-export * from './EnterButton';
-export * from './Main';
