--- client/src/recoil/cover.js
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const showPasswordModalAtom = atom({
-  key: 'showPasswordModal',
-  default: false,
-});
