--- client/src/recoil/index.js
@@ -1,3 +1,0 @@
-export * from './home';
-export * from './cover';
-export * from './paper-list';
