--- client/src/recoil/paper-list.js
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const showLetterModalAtom = atom({
-  key: 'showLetterModal',
-  default: false,
-});
