--- client/src/utils/dateConverter.js
@@ -1,5 +1,0 @@
-const dateConverter = (date) => {
-  return date.slice(0, 10);
-};
-
-export { dateConverter };
