+++ server/.gitignore
@@ -13,7 +13,6 @@ dist
 dist-ssr
 *.local
 .env
-json
 
 # Editor directories and files
 .vscode/*
